media-source: organize global constants
