--- projects/plugins/jetpack/extensions/shared/components/audio-player/constants.js
@@ -1,3 +1,0 @@
-export const STATE_PLAYING = 'is-playing';
-export const STATE_ERROR = 'is-error';
-export const STATE_PAUSED = 'is-paused';
